requirements.txt
discord
pytest
python-dotenv